refactor(dapp): split out cli utils
